Stop exceptions from having their package for DSE
