+++ SpiNNaker-data-specification/src/main/java/uk/ac/manchester/spinnaker/data_spec/ExecuteBreakInstruction.java
@@ -14,25 +14,13 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package uk.ac.manchester.spinnaker.data_spec.exceptions;
-
-import uk.ac.manchester.spinnaker.data_spec.Commands;
+package uk.ac.manchester.spinnaker.data_spec;
 
 /**
- * An exception that indicates that the value of the requested type is unknown.
+ * A special exception that indicates that a
+ * {@link uk.ac.manchester.spinnaker.data_spec.Commands#BREAK BREAK} was
+ * encountered.
  */
-public class UnknownTypeLengthException extends DataSpecificationException {
-	private static final long serialVersionUID = 8012093021275095441L;
-
-	/**
-	 * Create an instance.
-	 *
-	 * @param dataLen
-	 *            How long a set of data was requested.
-	 * @param command
-	 *            What command was being executed.
-	 */
-	public UnknownTypeLengthException(int dataLen, Commands command) {
-		super("Unknown data length " + dataLen + " during command " + command);
-	}
+public class ExecuteBreakInstruction extends DataSpecificationException {
+	private static final long serialVersionUID = -4902287652556707319L;
 }
