+++ SpiNNaker-data-specification/src/main/java/uk/ac/manchester/spinnaker/data_spec/RegionInUseException.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package uk.ac.manchester.spinnaker.data_spec.exceptions;
+package uk.ac.manchester.spinnaker.data_spec;
 
 /**
  * An exception that indicates that a region has already been allocated.
