+++ SpiNNaker-data-specification/src/test/java/uk/ac/manchester/spinnaker/data_spec/TestDataSpecExecutor.java
@@ -34,9 +34,6 @@ import java.nio.IntBuffer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import uk.ac.manchester.spinnaker.data_spec.exceptions.DataSpecificationException;
-import uk.ac.manchester.spinnaker.data_spec.exceptions.ExecuteBreakInstruction;
-
 public class TestDataSpecExecutor {
 
 	@Test
